fix merge error

+++ oneflow/core/operator/arg_where_op.cpp
@@ -76,7 +76,7 @@ class ArgWhereOp final : public Operator {
     size_t tmp_bytes = JUST(InferArgWhereTmpBufferSize(GetBlobDesc4BnInOp("in"),
                                                        op_conf().arg_where_conf().data_type()));
     OF_CHECK_GT(tmp_bytes, 0) << "tmp buffer bytes should be greater than zero";
-    tmp_desc->mut_shape() = Shape({tmp_bytes});
+    tmp_desc->mut_shape() = Shape({static_cast<int64_t>(tmp_bytes)});
     tmp_desc->set_data_type(DataType::kChar);
     return Maybe<void>::Ok();
   }
